+++ bundle/styles.css
@@ -2,9 +2,6 @@
 .container {
   display: flex;
   flex-direction: column;
-  max-width: 1600px;
-  margin-left: auto;
-  margin-right: auto;
 }
 
 /*###########################~~HEADER~~################################*/
@@ -50,8 +47,6 @@
   flex: 1.5;
   background-color: gold;
   padding: 5px;
-  padding-left: 15px;
-  min-width: 380px;
 }
 
 .overviewReviews {
@@ -62,8 +57,8 @@
 }
 
 .overviewNameAndCat {
-  /* display: flex; */
-  /* flex: 2; */
+  display: flex;
+  flex: 2;
   background-color: rgb(49, 245, 10);
   padding: 3px;
 }
@@ -96,10 +91,8 @@
   flex: 2.5;
   background-color: white;
   padding: 5px;
-  flex-wrap: wrap;
-  min-width: 680px;
-  /* min-height: 0px;
-  min-width: 0px; */
+  min-height: 0px;
+  min-width: 0px;
 }
 
 /*productDescription*/
@@ -123,69 +116,6 @@
   flex: 1.5;
   background-color: rgb(30, 255, 0);
 }
-
-/*thumbnails*/
-.selectStyle {
-  padding: 10px;
-  height: 55px;
-  width: 55px;
-  object-fit: cover;
-  border-radius: 50%;
-}
-
-/*buttons*/
-.selectSize {
-  display: block;
-  width: 65%;
-  margin-right: 5%;
-  padding-top: 15px;
-  padding-bottom: 15px;
-  max-height: 55px;
-  min-width: 225px;
-}
-
-.quantity {
-  display: block;
-  width: 30%;
-  padding-top: 15px;
-  padding-bottom: 15px;
-  max-height: 55px;
-  min-width: 104px;
-}
-
-.addToBag {
-  display: block;
-  width: 75%;
-  margin-right: 5%;
-  padding-top: 15px;
-  padding-bottom: 15px;
-  max-height: 55px;
-  min-width: 260px;
-}
-
-.star {
-  display: block;
-  width: 20%;
-  padding-top: 15px;
-  padding-bottom: 15px;
-  max-height: 55px;
-  min-width: 70px;
-}
-
-/* .category {
-  white-space: pre-wrap;
-} */
-
-/* .block {
-  display: block;
-  width: 100%;
-  border: none;
-  background-color: #04aa6d;
-  padding: 14px 28px;
-  font-size: 16px;
-  cursor: pointer;
-  text-align: center;
-} */
 
 /*###########################~~RELATED PRODUCTS AND COMPARISONS~~################################*/
 .rItemsCompare {
