+++ package.json
@@ -30,7 +30,6 @@
     "promise": "^8.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-image-file-resizer": "^0.4.7",
     "react-hooks-use-modal": "^2.1.0",
     "react-image-gallery": "^1.2.7",
     "underscore": "^1.13.2",
